dev_requirements.txt
-r ./requirements.txt
ipython==6.2.0
django-extensions==1.9.1
